move roa table code to table/ from server/
more appropriate.
